final version

+++ src/main/java/com/bloomberg/warehouse/repository/FxDealDAO.java
@@ -8,5 +8,5 @@ public interface FxDealDAO {
 
     void insertFxDeal(FxDeal fxDeal);
 
-    Optional<FxDeal> getFxDeal(int fxDealId);
+    Optional<FxDeal> getFxDeal(long fxDealId);
 }
