+++ src/main/java/com/bloomberg/warehouse/repository/FxDealRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FxDealRepository extends JpaRepository<FxDeal, Integer> {
+public interface FxDealRepository extends JpaRepository<FxDeal, Long> {
 
 }
